Revert ff9880689e0 and avoid meta key bindings
The prior attempt is flawed:

If you are in input mode, pressing ESC and a command may not lead to the
desired result. For example:

    Enter some text with several "the" (the the the)
    (while in normal mode)
    gg0
    /the
    a
    ^[
    n

In words: search for "the", enter input mode (append). Now hit ESC
and - as soon as possible - 'n' to search for the next "the". Instead of
jumping to the next match, a strange character is appended.

Another example:
(Make sure, ^[ is ESC, a single character)

    :let @a='a^[aexam^[aple^['
    @a

This should lead to

    example

but instead generates

    áexamáple